examples/ticker: Fix format specifier warning in low_ticker2
The format specifier expected size_t but uptime.tv_sec is of type
time_t. Cast uptime.tv_sec to int64_t and use %" PRId64 " to match the argument
type and avoid compiler warnings.

Related to #9

+++ ticker/low_ticker2/init.c
@@ -29,8 +29,8 @@ rtems_task Test_task(
       printk( "*** END OF LOW MEMORY CLOCK TICK TEST (delay) ***\n" );
       rtems_shutdown_executive( 0 );
     }
-    printk( "TA%" PRIuPTR " - rtems_clock_uptime - %zd:%ld\n", 
-      task_index, uptime.tv_sec, uptime.tv_nsec 
+    printk( "TA%" PRIuPTR " - rtems_clock_uptime - %" PRId64 ":%ld\n", 
+      task_index,(int64_t)uptime.tv_sec, uptime.tv_nsec 
     );
   }
 }
